% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.clipTranscript}
\alias{.clipTranscript}
\title{Clip Transcript to given length}
\usage{
.clipTranscript(grlExons, maxTxLength, txEnd, BPPARAM)
}
\arguments{
\item{grlExons}{a \code{CompressedGRangesList} object}

\item{maxTxLength}{a positive integer}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether
and how the method should be parallelized.}
}
\value{
the clipped \code{GRanges} object
}
\description{
Internal function for operating on a \code{GRangesList}, where ranges
represent exons in a transcript. This is designed to be used with a
\code{GRangesList} object.
}
