% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.clipTranscript_modded2}
\alias{.clipTranscript_modded2}
\title{Clip Transcript to Given Length - Modded 2 by Guillermo R.}
\usage{
.clipTranscript_modded2(grlExons, maxTxLength, BPPARAM)
}
\arguments{
\item{maxTxLength}{a positive integer}

\item{gr}{a \code{GRanges} object}
}
\value{
the clipped \code{GRanges} object
}
\description{
Internal function for operating on individual \code{GRanges}, where ranges
represent exons in a transcript. This is designed to be used in an
\code{*apply} function over a \code{GRangesList} object.
}
