% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.generateGeneRanges}
\alias{.generateGeneRanges}
\title{Generates the genes ranges from a \code{GRanges}}
\usage{
.generateGeneRanges(grExons)
}
\arguments{
\item{grlC_clipped}{a \code{GRanges} of truncated exons.}
}
\value{
\code{GRanges} with genes defined from the truncated exons
}
\description{
Generates the genes ranges from a \code{GRanges}
}
