% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.generateTranscriptRanges}
\alias{.generateTranscriptRanges}
\title{Generates the transcript ranges from a \code{GRanges}}
\usage{
.generateTranscriptRanges(grExons)
}
\arguments{
\item{grlC_clipped}{a \code{GRanges} of truncated exons.}
}
\value{
\code{GRanges} with transcripts defined from the truncated exons
}
\description{
Generates the transcript ranges from a \code{GRanges}
}
