% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.pruneInconsistentStrand}
\alias{.pruneInconsistentStrand}
\title{Prune GRanges of transcripts with inconsistent strand}
\usage{
.pruneInconsistentStrand(exons_df)
}
\arguments{
\item{exons_df}{a \code{tibble} of exons with \code{strand} and
\code{transcript_id} fields.}
}
\value{
\code{tibble} with inconsistent transcripts removed
}
\description{
Prune GRanges of transcripts with inconsistent strand
}
