% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeTxDbSafe}
\alias{makeTxDbSafe}
\title{Compatibility wrapper for \code{makeTxDbFromGRanges}}
\usage{
makeTxDbSafe(...)
}
\arguments{
\item{...}{Arguments passed directly to
\code{makeTxDbFromGRanges()} from either \pkg{txdbmaker} or
\pkg{GenomicFeatures}.}
}
\value{
A \link[GenomicFeatures:TxDb-class]{TxDb} object constructed
from the provided \code{GRanges} object(s), consistent with
the behavior of \code{makeTxDbFromGRanges()}.
}
\description{
This internal utility provides a version-safe interface to the
\code{makeTxDbFromGRanges()} function, which has been moved from the
\pkg{GenomicFeatures} package to the \pkg{txdbmaker} package in
Bioconductor (>= 3.19, GenomicFeatures >= 1.61.1).
}
\details{
The function checks at runtime which package provides
\code{makeTxDbFromGRanges()} and dispatches the call accordingly.

If \pkg{txdbmaker} is installed, the function calls
\code{txdbmaker::makeTxDbFromGRanges()}.
If not, and if \pkg{GenomicFeatures} provides the function (in versions
\eqn{\leq 1.61.1}), it will call
\code{GenomicFeatures::makeTxDbFromGRanges()} instead.
If neither package is available, an error is raised.
}
