% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{truncateTxome}
\alias{truncateTxome}
\alias{truncate3primeTxome}
\alias{truncate5primeTxome}
\alias{truncate3primeTxome,TxDb-method}
\alias{truncate5primeTxome,TxDb-method}
\alias{truncateTxome,TxDb-method}
\title{Truncate Transcriptome}
\usage{
truncateTxome(
  txdb,
  maxTxLength = 500,
  txEnd = "3prime",
  overlap_path = "",
  BPPARAM = bpparam()
)

truncate3primeTxome(
  txdb,
  maxTxLength = 500,
  overlap_path = "",
  BPPARAM = bpparam(),
  quiet = F
)

truncate5primeTxome(
  txdb,
  maxTxLength = 300,
  overlap_path = "",
  BPPARAM = bpparam(),
  quiet = F
)

\S4method{truncate3primeTxome}{TxDb}(txdb, maxTxLength = 500, BPPARAM = bpparam(), quiet = F)

\S4method{truncate5primeTxome}{TxDb}(txdb, maxTxLength = 300, BPPARAM = bpparam(), quiet = F)

\S4method{truncateTxome}{TxDb}(
  txdb,
  maxTxLength = 500,
  txEnd = "3prime",
  overlap_path = "",
  BPPARAM = bpparam()
)
}
\arguments{
\item{txdb}{a \code{TxDb} object}

\item{maxTxLength}{the maximum length of transcripts}

\item{txEnd}{transcript end to truncate}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether
and how the method should be parallelized.}

\item{quiet}{silent output messages}

\item{overlapPath}{path to store the transcripts removed after truncation due
to overlaps. Leave empty to not store any results.}
}
\value{
a \code{TxDb} object
}
\description{
Truncates the provided transcriptome
}
\examples{
library(TxDb.Scerevisiae.UCSC.sacCer3.sgdGene)

## load annotation
txdb <- TxDb.Scerevisiae.UCSC.sacCer3.sgdGene

## restrict to 'chrI' transcripts
seqlevels(txdb) <- c("chrI")

## last 500 nts per tx
txdb_w500 <- truncate3primeTxome(txdb)
txdb_w500

## last 100 nts per tx
txdb_w100 <- truncate3primeTxome(txdb, maxTxLength=100)
txdb_w100
}
